refactor: 콘솔 패키지 구조 변경
